// src/api/generated/models/questionModel.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * Tg-Api 
 * OpenAPI spec version: 1.0.0
 */
import type { QuestionModelCodeSnippet } from './questionModelCodeSnippet';
import type { QuestionResponseModel } from './questionResponseModel';

export interface QuestionModel {
  id: string;
  question: string;
  /** @nullable */
  question_pt: string | null;
  /** @nullable */
  question_en: string | null;
  translation_status: string;
  difficulty_level: number;
  /** @nullable */
  language_id: string | null;
  category: string;
  ideal_time_to_solve: number;
  code_snippet: QuestionModelCodeSnippet;
  is_multiple_choice: boolean;
  seniority_level: string;
  responses: QuestionResponseModel[];
}