// src/api/generated/models/questionResponseModel.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * Tg-Api 
 * OpenAPI spec version: 1.0.0
 */
import type { QuestionResponseModelCodeSnippet } from './questionResponseModelCodeSnippet';

export interface QuestionResponseModel {
  id: string;
  question_id: string;
  response: string;
  /** @nullable */
  response_pt: string | null;
  /** @nullable */
  response_en: string | null;
  translation_status: string;
  is_correct: boolean;
  code_snippet: QuestionResponseModelCodeSnippet;
}